+++ examples/compareDataSets.c
@@ -5,7 +5,7 @@
 
 int main(int argc, char * argv[])
 {	
-    int r5=0,r4=0,r3=0,r2=0,r1=0;
+    size_t r5=0,r4=0,r3=0,r2=0,r1=0;
     char oriFilePath[640], decFilePath[640];
     char *cfgFile, *compressionCase, *varName;
 
@@ -37,7 +37,7 @@ int main(int argc, char * argv[])
     printf("decFilePath=%s\n", decFilePath);
     ZC_Init(cfgFile);
 
-    int nbEle1, nbEle2;
+    size_t nbEle1, nbEle2;
     float *data1 = ZC_readFloatData(oriFilePath, &nbEle1);
     float *data2 = ZC_readFloatData(decFilePath, &nbEle2);
